refactor(cli): isolate export helpers
